requirements.txt
TWILIO_ACCOUNT_SID : AC0af45dfbf51ccb9be2517a98351c61fc
TWILIO_AUTH_TOKEN : aed0fbc2ae6f7b3f15c14b3c6a196cf7
TWILIO_PHONE_NUMBER : +18142125558
TELEGRAM_BOT_TOKEN : 8419741449:AAGsbPs78s1YQnUK5enbL2WQMma4zBQKJVY